Moved Python example
