+++ examples/python/requirements.txt
@@ -1,0 +1,1 @@
+pyserial == 3.5
