+++ examples/python/serial_client.py
