--- python/requirements.txt
@@ -1,2 +1,0 @@
-keyboard==0.13.5
-numpy==1.26.0
